requirements.txt
langchain>=0.1.0
langchain-community>=0.0.21
langchain-openai>=0.0.5
python-dotenv>=1.0.0
openai>=1.12.0
chromadb>=0.4.22
tiktoken>=0.5.2
gitpython>=3.1.41
